backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
feedparser==6.0.11
supabase==2.9.0
python-dotenv==1.0.1
APScheduler==3.10.4
